refactor: Move CLI, cleanup debug statements, format code
