+++ lib/ai/mock-functions.ts
